+++ packages/markdown-overrides/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.9.8](https://github.com/Codecademy/gamut/compare/@codecademy/markdown-overrides@0.9.7...@codecademy/markdown-overrides@0.9.8) (2022-07-29)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.9.7](https://github.com/Codecademy/gamut/compare/@codecademy/markdown-overrides@0.9.6...@codecademy/markdown-overrides@0.9.7) (2022-07-29)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
+### [0.9.6](https://github.com/Codecademy/gamut/compare/@codecademy/markdown-overrides@0.9.5...@codecademy/markdown-overrides@0.9.6) (2022-07-27)
+
+**Note:** Version bump only for package @codecademy/markdown-overrides
+
+
+
+
+
 ### [0.9.5](https://github.com/Codecademy/gamut/compare/@codecademy/markdown-overrides@0.9.4...@codecademy/markdown-overrides@0.9.5) (2022-07-19)
 
 **Note:** Version bump only for package @codecademy/markdown-overrides